% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices.R
\name{choices}
\alias{choices}
\title{A question to pick among several choices}
\usage{
choices(
  prompt = "The question",
  bad = list(Blue = "Blue is too bright!", Red = "Red is too distracting"),
  good = list(Green = "Yuck!!"),
  points = 5,
  need_all = length(good) > 1,
  allow_retry = TRUE
)
}
\description{
A question to pick among several choices
}
