% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_liner.R
\name{one_liner}
\alias{one_liner}
\title{Grade one-line expressions}
\usage{
one_liner(user, soln, correct = "Good", incorrect = "Bad")
}
\arguments{
\item{user}{user code as a character string}

\item{soln}{solution code as a character string}
}
\description{
One line expressions consist of just a single expression.
}
