% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_check_inputs.R
\name{save_check_inputs}
\alias{save_check_inputs}
\title{Save learnr  exercise checking inputs to a file}
\usage{
save_check_inputs(...)
}
\description{
Save learnr  exercise checking inputs to a file
}
