% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yesno.R
\name{yesno}
\alias{yesno}
\title{Yes/No questions}
\usage{
yesno(
  prompt = "Is the moon made of blue cheese?",
  istrue = TRUE,
  right = "Good job",
  wrong = "Not so good!",
  points = 5
)
}
\arguments{
\item{prompt}{Character  string asking the true or false question.}

\item{istrue}{logical containing answer to the question \code{TRUE} or \code{FALSE}}

\item{right}{character string to give for correct answer}

\item{wrong}{character string  to give for wrong answer}

\item{points}{Number: How many points is this  question worth. \code{NULL}  if
you don't want to set a number of points.}
}
\description{
Yes/No questions
}
\details{
The words used for true and false and the words used to indicate
a correct answer are selected at random. But they are determined when
the document is compiled, so they will  be random only  from one question to another
not from multiple responses to the same question.
}
